Add left+right key press navigation!
You can now navigate across all criteria with the left and right keys.
The only issue is that you can't hold the key and have it go till the end.
You currently have to press they key every time.

It won't go under 0 or over the length of the criteria array