Add left/right navigation on the last criteria for report
Now if you go to the right after the last item, you go to the report page.
If you go left from there, you go back to the last criteria.

Now, we could technically remove up/down navigation. Although that would
mean that if you are near the middle of the criteria and you want to see
the current report, you would have to navigate all the way to the right to
see it instead of being able to just use the up/down nav